fix: refactor markdown paste, reprioritize paste command listeners
