+++ lib/lexical/exts/markdown.js
@@ -10,6 +10,7 @@ import {
   defineExtension,
   $getSelection,
   $isRangeSelection,
+  $getRoot,
   PASTE_COMMAND,
   COPY_COMMAND,
   CUT_COMMAND,
@@ -18,21 +19,49 @@ import {
 import { registerRichText } from '@lexical/rich-text'
 import { mergeRegister, objectKlassEquals } from '@lexical/utils'
 import { $insertText } from '@/lib/lexical/utils'
+import { $lexicalToMarkdown } from '@/lib/lexical/utils/mdast'
+import { withDisposableBridge } from '@/components/editor/hooks/use-headless-bridge'
+
+/** redirects a paste event to a disposable rich text bridge,
+ *  and exports the resulting EditorState as markdown via MDAST */
+const getMarkdownFromPaste = withDisposableBridge((bridge, event) => {
+  try {
+    bridge.update(() => {
+      const root = $getRoot()
+      root.clear()
+      root.selectEnd()
+    }, { discrete: true })
+
+    bridge.dispatchCommand(PASTE_COMMAND, event)
+
+    let markdown = null
+    bridge.update(() => {
+      markdown = $lexicalToMarkdown(true)
+    }, { discrete: true })
+
+    return markdown
+  } catch {
+    return null
+  }
+}, { name: 'sn-markdown-paste-bridge' })
 
 function onPasteForMarkdown (event, editor) {
   event.preventDefault()
-  editor.update(() => {
-    const selection = $getSelection()
-    // anything else is managed by other Lexical commands
-    const clipboardData = objectKlassEquals(event, window.ClipboardEvent) ? event.clipboardData : null
-    if (clipboardData != null && selection !== null) {
-      // we'll just get the plain text data from the clipboard
-      const text = clipboardData.getData('text/plain') || clipboardData.getData('text/uri-list')
-      if (text != null) {
-        $insertText(text, true)
-      }
-    }
-  }, { tag: PASTE_TAG })
+
+  const clipboardData = objectKlassEquals(event, window.ClipboardEvent)
+    ? event.clipboardData
+    : null
+  if (!clipboardData) return
+
+  // check for rich content (lexical or HTML)
+  const hasRichContent = clipboardData.getData('application/x-lexical-editor') || clipboardData.getData('text/html')
+  // if we have rich content, get the markdown from the paste
+  // otherwise, use plain text
+  const text = (hasRichContent && getMarkdownFromPaste(event)) ||
+    clipboardData.getData('text/plain') || clipboardData.getData('text/uri-list')
+
+  if (!text) return
+  editor.update(() => $insertText(text, true), { tag: PASTE_TAG })
 }
 
 /** rich text extension that handles plain text only */
@@ -69,11 +98,10 @@ export const MarkdownTextExtension = defineExtension({
         PASTE_COMMAND,
         (e) => {
           const selection = $getSelection()
-          if (selection !== null) {
-            onPasteForMarkdown(e, editor)
-            return true
-          }
-          return false
+          if (!$isRangeSelection(selection)) return false
+
+          onPasteForMarkdown(e, editor)
+          return true
         },
         COMMAND_PRIORITY_NORMAL
       ),
